fix segfault in Connection::getFunction()
the segfault ocurred during testing:
* lookup an existing function (eg RFC_PING)
* then pass a non-string parameter (eg []) to getFunction
-> segfault in zend_string_init() (seemed random)

the root cause for this mess was the RemoteFunction::$name property.
it was assigned the function_name zend_string from the method arguments,
not a copy. it seems the zend_string got released/gc'd/whatever...
now we copy the function_name for the $name property and all seems good.
it might have also worked if we used add_property_string() with
ZSTR_VAL(function_name), as that would have create a new zend_string for
the property.

this might be the reason for the segfault in issue #11