improve parameter handling
the status of the parameters is now saved in a HashTable and the function
handle is now created in invoke(), not getFunction(). this enables reuse
of the RemoteFunction object.
all parameters are active by default, so RfcSetParameterActive is only
called if the user deactivated a parameter.
we cant call it for every parameter and set it to active as this causes
really strange errors and segfaults if the FM throws an ABAP exception
(eg use RFC_RAISE_ERROR test 043 to reproduce - when forcing parameters
active, the test segfaults because the exception is ignored and the code
tries to fetch a return value for CSTRING). there might be a better
solution, but for now it works.

refs #15