simplify test handling
previously one had to define a RUN_ONLINE_TESTS=1 environment
variable for make test and a configuration file to run all tests.
as not all installations have E in their variables order, one had
to change the PHP configuration to allow for all tests to run.

to make this easier, now it suffices to provide a configuration
file in the tests directory. if one is detected, all tests are
enabled when running make test.