improve exception messages for TABLE parameters
previously, if an exception occurred when setting table rows
or getting table lines, the exception message would not name the
parameter in question. to ease debugging, the parameter name is
now included in all exception messages (it was also missing in
a NUM parameter exception).

so instead of eg.
  "Failed to set TABLE row, expected array"
the message is now
  "Failed to set TABLE row, expected array for parameter ORDER_PARTNERS"

refs #26