allow parameters to be passed as references
parameters passed as reference are dereferenced before checking
the types. previously, passing eg an array by reference as parameter
for a TABLE would fail because the type check would return IS_REF
instead of IS_ARRAY.

now every setter checks if the passed parameter is a reference first.
if yes, it is dereferenced before checking the actual type requirements.

refs #26
closes #27