add method to fetch a RemoteFunction's description
RemoteFunction::getFunctionDescription() returns an array describing
the function module's interface.

The array is indexed with the parameter names and each parameter
description contains the keys 'name', 'type', 'direction',
'description', 'ucLength', 'nucLength', 'decimals', 'optional' and
'default'.

For structures and tables, an additional key 'typedef' contains an array
with the type's field descriptions. The type description is indexed with
the field name and contains the keys 'name', 'type', 'ucLength',
'ucOffset', 'nucLength', 'nucOffset' and 'decimals'.

refs #79