Add support for retrieving SSO ticket from connection
Connection::getSSOTicket() can be used to retrieve a SSO ticket from an
existing connection.

This is only supported with SDK versions 7.50.1 and up. The connection
must be opened with parameter `GETSSO2=1` and the profile parameter
`login/create_sso2_ticket` must be set to a value different than 0 in
the backend.

refs #34