Allow mocking RemoteFunction
Remove the ZEND_ACC_FINAL from the RemoteFunction class entry to make it
mockable.

Also fix a segfault in sapnwrfc_function_object_free() when freeing the
function name.  This happens when manually constructing a RemoteFunction for
tests and no name is set (which usually happens in Connection::getFunction()).

Also fix an issue when calling zend_hash_destroy() on the parameter_status of a
manually constructed RemoteFunction object. Move the zend_hash_init() to
sapnwrfc_function_object_create(). This fixes a a failing assertion in
_zend_is_inconsistent (ht=0x.... is being cleaned).

refs #133