Fix: don't call RfcDestroyFunctionDesc in sapnwrfc_function_object_free
RfcDestroyFunctionDesc deletes the function description and releases the
allocated resources. This causes an issue when looking up a function,
unsetting the RemoteFunction object, looking up the same function and
invoking it again.

This either leads to a CALL_FUNCTION_NOT_FOUND error or a segfault
(depending on whether the function description cache is cleard in
between).

Simply unset the function_desc_handle and keep the function description
in the cache. If a user wants to clear the cache, they can always do so
manually.

Also update docs and clarify the $repositoryId parameter of
clearFunctionDescCache().