Throw if reading connection attributes fails
In sapnwrfc_open_connection(), return after throwing exception. There is
not point in going on if we don't get a valid handle.

Also check RfcGetConnectionAttributes error code and throw if fetching
the attributes fails.

This fixes a "warning: the comparison will always evaluate as ‘true’ for
the address of ‘sysId’ will never be NULL [-Waddress]" as well.